refactor: move extract_api_definition.py into get_api_brain_definition_as_json_schema.py
